backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
archspec==0.2.3
bcrypt==3.2.0
boltons==24.0.0
Brotli==1.1.0
cachetools==5.3.3
catkin-pkg==1.0.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
docutils==0.21.2
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
google-ai-generativelanguage==0.6.5
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.6
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.5
packaging==24.0
pip==24.0
platformdirs==4.2.0
pluggy==1.4.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycosat==0.6.6
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
truststore==0.8.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
wheel==0.43.0
zstandard==0.22.0
sqlalchemy>=1.4,<2.0
mysqlclient>=2.0.0,<3.0.0
passlib>=1.7.4,<2.0.0
PyJWT>=2.0.0,<3.0.0
pymupdf>=1.19.6,<2.0.0
python-pptx>=0.6.21,<0.7.0
python-docx>=0.8.11,<0.9.0
jsonpickle
